requirements.txt
google-adk>=1.2.1
supabase>=2.15.2
langchain>=0.3.25
langgraph>=0.4.8
langchain-google-genai>=2.1.5
python-dotenv>=1.1.0
fastapi>=0.115.12
uvicorn>=0.34.3
websockets>=14.2
asyncio-mqtt>=0.16.2
pydantic>=2.11.5
httpx>=0.28.1
aiofiles>=24.1.0
pydantic-settings>=2.9.1
python-multipart>=0.0.20
sse-starlette>=2.3.6 